# perf(pnp): enable strict mode in `.pnp.cjs`