**What's the problem this PR addresses?**

The `.pnp.cjs` file doesn't use strict mode because `Emscripten libraries aren't strict mode` but that isn't the case anymore.

**How did you fix it?**

Enabled strict mode.

**Benchmark result**

Doesn't do much for performance but either way:
```console
$ hyperfine "node -r ./.pnp-before.cjs -e 1" "node -r ./.pnp-after.cjs -e 1"
Benchmark 1: node -r ./.pnp-before.cjs -e 1
  Time (mean ± σ):      91.7 ms ±   2.7 ms    [User: 1.9 ms, System: 3.9 ms]
  Range (min … max):    89.6 ms …  99.7 ms    32 runs

Benchmark 2: node -r ./.pnp-after.cjs -e 1
  Time (mean ± σ):      91.0 ms ±   1.1 ms    [User: 0.5 ms, System: 1.6 ms]
  Range (min … max):    89.7 ms …  94.6 ms    31 runs

Summary
  'node -r ./.pnp-after.cjs -e 1' ran
    1.01 ± 0.03 times faster than 'node -r ./.pnp-before.cjs -e 1'
```


**Checklist**
- [x] I have read the [Contributing Guide](https://yarnpkg.com/advanced/contributing).
- [x] I have set the packages that need to be released for my changes to be effective.
- [x] I will check that all automated PR checks pass before the PR gets reviewed.